backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1

# Neo4j + Vector Store
neo4j==5.15.0
weaviate-client==4.4.0
sentence-transformers>=2.5.0
numpy==1.24.3
pandas==2.0.3

# LangGraph + LangChain
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-neo4j>=0.0.1
langchain-weaviate>=0.0.1

# Celery for async task processing
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

